feat: add readiness checks for unknown resources (#6)
* feat: add readiness checks for unknown resources

Signed-off-by: Dustin Scott <dustin.scott18@gmail.com>

* chore: fix typo in ephemeral

Signed-off-by: Dustin Scott <dustin.scott18@gmail.com>

* fix: stop processing get watches

This commit stops processing the get watches loop if we found a watch already
stored for a particular object.  This is to slightly improve efficiency when
you have lots of objects and a match is found early in the loop.

Signed-off-by: Dustin Scott <dustin.scott18@gmail.com>

* refactor: change crud methods to include reusable logger and efficient returns

Added quick return for the update method.  This allows the code
to stop processing and return immediately.  Also added re-usable
logger methods for CRUD operations to slim down code a bit.

Signed-off-by: Dustin Scott <dustin.scott18@gmail.com>

* refactor: export and re-use MessageFor method

This allows the method to be re-used across other packages.  Also implemented the
MessageFor use across other package which use this same logic.

Signed-off-by: Dustin Scott <dustin.scott18@gmail.com>

* refactor: make for loop more efficient

Reduce need to copy large objects in memory for each for loop.  This is helpful because
the predicates are frequently called, so we want to reduce memory wherever we can.

Signed-off-by: Dustin Scott <dustin.scott18@gmail.com>

---------

Signed-off-by: Dustin Scott <dustin.scott18@gmail.com>